refactor: 레이어별 패키지 구조 분리
