+++ src/main/java/roomescape/controller/ReservationTimeController.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.controller;
 
 import java.util.List;
 import org.springframework.http.ResponseEntity;
@@ -8,6 +8,8 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
+import roomescape.domain.ReservationTime;
+import roomescape.dao.ReservationTimeDao;
 
 @RestController
 public class ReservationTimeController {
