+++ src/main/java/roomescape/dao/ReservationTimeDao.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.dao;
 
 import java.sql.PreparedStatement;
 import java.util.List;
@@ -6,6 +6,7 @@ import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
+import roomescape.domain.ReservationTime;
 
 @Repository
 public class ReservationTimeDao {
