+++ src/main/java/roomescape/dto/ReservationRequest.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.dto;
 
 public class ReservationRequest {
     private final String name;
